requirements.txt
pandas
defopt~=5.1.0
yfinance==0.1.66
mysql-connector-python==8.0.27

